tests/wpt/mozilla/tests/bluetooth/getDescriptors/descriptors-found-with-uuid.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(generic_access.name))
    .then(service => service.getCharacteristic(device_name.name))
    .then(characteristic => {
        return Promise.all([
            characteristic.getDescriptors(number_of_digitals.alias),
            characteristic.getDescriptors(number_of_digitals.name),
            characteristic.getDescriptors(number_of_digitals.uuid)])
        .then(descriptors_arrays => {
            descriptors_arrays.forEach(descriptors => {
                assert_equals(descriptors.length, 2);
                assert_equals(descriptors[0].uuid, number_of_digitals.uuid);
                assert_equals(descriptors[1].uuid, number_of_digitals.uuid);
            });
        });
    });
}, 'Find descriptors with UUID in service.');
</script>
